% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odv_contours.R
\name{ggplot_add.divaodv_contours_remove}
\alias{ggplot_add.divaodv_contours_remove}
\title{Remove ODV contour layers from a ggplot}
\usage{
\method{ggplot_add}{divaodv_contours_remove}(object, plot, object_name)
}
\arguments{
\item{object}{A \code{divaodv_contours_remove}.}

\item{plot}{The ggplot being added to.}

\item{object_name}{Name of the object, supplied by ggplot2.}
}
\value{
The modified ggplot.
}
\description{
Remove ODV contour layers from a ggplot
}
