% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_config.R
\name{print.diva_variable_config}
\alias{print.diva_variable_config}
\title{Print method for diva_variable_config}
\usage{
\method{print}{diva_variable_config}(x, ...)
}
\arguments{
\item{x}{A \code{diva_variable_config} object.}

\item{...}{Ignored.}
}
\description{
Print method for diva_variable_config
}
