% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odv_contours.R
\name{print.divaodv_contours}
\alias{print.divaodv_contours}
\title{Print a deferred ODV contour specification}
\usage{
\method{print}{divaodv_contours}(x, ...)
}
\arguments{
\item{x}{A \code{divaodv_contours}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a deferred ODV contour specification
}
