% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odv_contours.R
\name{print.divaodv_contours_remove}
\alias{print.divaodv_contours_remove}
\title{Print a deferred ODV contour removal}
\usage{
\method{print}{divaodv_contours_remove}(x, ...)
}
\arguments{
\item{x}{A \code{divaodv_contours_remove}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a deferred ODV contour removal
}
