% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_odv.R
\name{print.divaodv_deferred_scale}
\alias{print.divaodv_deferred_scale}
\title{Print a deferred ODV fill scale}
\usage{
\method{print}{divaodv_deferred_scale}(x, ...)
}
\arguments{
\item{x}{A \code{divaodv_deferred_scale}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a deferred ODV fill scale
}
